WIP: Unified into single Processor; however, we are still getting a separate adapter class for the second round.
We need some type of marker mechanism in order to prevent generating a second (meaningless) adapter in the second round.
