+++ extensions/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,2 +1,1 @@
 org.mapstruct.extensions.spring.converter.ConverterMapperProcessor
-org.mapstruct.extensions.spring.converter.DelegatingConverterProcessor
