+++ extensions/src/test/java/org/mapstruct/extensions/spring/converter/ConverterMapperProcessorTest.java
@@ -1,5 +1,7 @@
 package org.mapstruct.extensions.spring.converter;
 
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.RetentionPolicy.SOURCE;
 import static javax.lang.model.element.Modifier.*;
 import static org.assertj.core.api.BDDAssertions.then;
 import static org.mockito.ArgumentMatchers.any;
@@ -7,6 +9,8 @@ import static org.mockito.ArgumentMatchers.any;
 import com.squareup.javapoet.*;
 import java.io.IOException;
 import java.io.Writer;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
 import java.lang.ref.WeakReference;
 import java.time.Clock;
 import java.util.Locale;
@@ -14,6 +18,7 @@ import javax.tools.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mapstruct.Mapper;
+import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.mapstruct.extensions.spring.ExternalConversion;
 import org.mockito.*;
 import org.mockito.junit.jupiter.MockitoExtension;
@@ -21,7 +26,6 @@ import org.springframework.core.convert.converter.Converter;
 
 @ExtendWith(MockitoExtension.class)
 class ConverterMapperProcessorTest extends AbstractProcessorTest {
-
   @Spy
   private final ConversionServiceAdapterGenerator adapterGenerator =
       new ConversionServiceAdapterGenerator(Clock.systemUTC());
@@ -33,6 +37,10 @@ class ConverterMapperProcessorTest extends AbstractProcessorTest {
   @Spy
   private final ConverterScansGenerator converterScansGenerator =
       new ConverterScansGenerator(Clock.systemUTC());
+
+  @Spy
+  private final DelegatingConverterGenerator delegatingConverterGenerator =
+      new DelegatingConverterGenerator(Clock.systemUTC());
 
   @Spy
   private final ConverterRegistrationConfigurationGenerator
@@ -223,5 +231,60 @@ class ConverterMapperProcessorTest extends AbstractProcessorTest {
             new FromToMapping()
                 .source(ClassName.get(String.class))
                 .target(ClassName.get(Locale.class)));
+  }
+
+  @Test
+  void shouldCompileMapperWithDelegatingConverterAnnotation() throws IOException {
+    final var delegatingConverterAnnotationTypeSpec =
+        TypeSpec.annotationBuilder(ClassName.get(DelegatingConverter.class))
+            .addAnnotation(
+                AnnotationSpec.builder(Target.class).addMember("value", "$L", METHOD).build())
+            .addAnnotation(
+                AnnotationSpec.builder(Retention.class).addMember("value", "$L", SOURCE).build())
+            .addModifiers(PUBLIC)
+            .build();
+    final var delegatingConverterFile =
+        JavaFile.builder(
+                DelegatingConverter.class.getPackageName(), delegatingConverterAnnotationTypeSpec)
+            .addStaticImport(METHOD)
+            .addStaticImport(SOURCE)
+            .build();
+    final var mapperTypeSpec =
+        TypeSpec.classBuilder("CarMapper")
+            .addAnnotation(Mapper.class)
+            .addModifiers(PUBLIC, ABSTRACT)
+            .addSuperinterface(
+                ParameterizedTypeName.get(
+                    ClassName.get(Converter.class), CAR_CLASS_NAME, CAR_DTO_CLASS_NAME))
+            .addMethod(
+                MethodSpec.methodBuilder("convert")
+                    .addModifiers(PUBLIC, ABSTRACT)
+                    .addParameter(CAR_CLASS_NAME, "car")
+                    .returns(CAR_DTO_CLASS_NAME)
+                    .build())
+            .addMethod(
+                MethodSpec.methodBuilder("invert")
+                    .addAnnotation(DelegatingConverter.class)
+                    .addModifiers(PUBLIC, ABSTRACT)
+                    .addParameter(CAR_DTO_CLASS_NAME, "carDto")
+                    .returns(CAR_CLASS_NAME)
+                    .build())
+            .build();
+    final var mapperFile = JavaFile.builder(PACKAGE_NAME, mapperTypeSpec).build();
+    final var autowiredTypeSpec =
+        TypeSpec.annotationBuilder(
+                ClassName.get("org.springframework.beans.factory.annotation", "Autowired"))
+            .addModifiers(PUBLIC)
+            .build();
+    final var autowiredFile =
+        JavaFile.builder("org.springframework.beans.factory.annotation", autowiredTypeSpec).build();
+
+    final var compileResult =
+        compile(
+            delegatingConverterFile.toJavaFileObject(),
+            mapperFile.toJavaFileObject(),
+            autowiredFile.toJavaFileObject());
+
+    then(compileResult).isTrue();
   }
 }
